+++ docs/.vitepress/theme/index.ts
@@ -1,3 +1,12 @@
+/*
+ * @Author: weisheng
+ * @Date: 2024-10-12 22:09:33
+ * @LastEditTime: 2024-11-09 23:03:07
+ * @LastEditors: weisheng
+ * @Description: 
+ * @FilePath: /wot-design-uni/docs/.vitepress/theme/index.ts
+ * 记得注释
+ */
 import { h } from 'vue'
 import Theme from 'vitepress/theme'
 import './styles/vars.css'
@@ -10,7 +19,10 @@ import CustomFooter from './components/CustomFooter.vue'
 import SvgImage from './components/SvgImage.vue'
 import HomeStar from './components/HomeStar.vue'
 import frame from './components/frame.vue'
-import { ElTag, ElLink } from 'element-plus'
+import ExternalLink from './components/ExternalLink.vue'
+import ElementPlus from 'element-plus'
+import 'element-plus/dist/index.css'
+
 
 export default {
   ...Theme,
@@ -25,7 +37,7 @@ export default {
   enhanceApp({ app }) {
     app.component('SvgImage', SvgImage)
     app.component('frame', frame)
-    app.component('ElTag', ElTag)
-    app.component('ElLink', ElLink)
+    app.component('ExternalLink',ExternalLink)
+    app.use(ElementPlus)
   },
 }