+++ src/pages/button/Index.vue
@@ -2,7 +2,7 @@
   <page-wraper>
     <view>
       <demo-block title="基本用法">
-        <wd-button open-type="getUserInfo" @getuserinfo="handleGetuserinfo">主要按钮</wd-button>
+        <wd-button>主要按钮</wd-button>
         <wd-button type="success">成功按钮</wd-button>
         <wd-button type="info">信息按钮</wd-button>
         <wd-button type="warning">警告按钮</wd-button>