+++ src/pages/form/demo2.vue
@@ -3,14 +3,14 @@
     <wd-form ref="form" :model="model" :reset-on-change="false">
       <wd-cell-group border>
         <wd-input
-          label="用户名"
+          label="歪比巴卜"
           label-width="100px"
           prop="name"
           clearable
           v-model="model.name"
-          placeholder="请输入用户名"
+          placeholder="请输入歪比巴卜"
           @blur="handleBlur('name')"
-          :rules="[{ required: true, message: '请填写用户名' }]"
+          :rules="[{ required: true, message: '请填写歪比巴卜' }]"
         />
         <wd-input
           label="玛卡巴卡单号"